refactor(api): mount application routes and rename hooks
