+++ packages/core/src/routes/types.ts
@@ -4,7 +4,7 @@ import type Router from 'koa-router';
 import type { WithLogContext } from '#src/middleware/koa-audit-log.js';
 import type { WithAuthContext } from '#src/middleware/koa-auth/index.js';
 import type { WithI18nContext } from '#src/middleware/koa-i18next.js';
-import { type WithHookContext } from '#src/middleware/koa-management-api-hooks.js';
+import { type WithHookContext } from '#src/middleware/koa-api-hooks.js';
 import type TenantContext from '#src/tenants/TenantContext.js';
 
 import { type WithAccountCenterContext } from './account/middlewares/koa-account-center.js';
